+++ Demo/JSONPreviewDemo/Example/Controller/HideLineNumberExampleViewController.swift
@@ -46,7 +46,7 @@ private extension HideLineNumberExampleViewController {
         print("will display json")
         let start = Date().timeIntervalSince1970
         
-        previewView.preview(json, initialState: .folded) {
+        previewView.preview(json, initialState: .folded) { _ in
             let end = Date().timeIntervalSince1970
             print("did display json at: \(end - start)")
         }