+++ Demo/JSONPreviewDemo/Example/Controller/InitialFoldedExampleViewController.swift
@@ -42,7 +42,7 @@ private extension InitialFoldedExampleViewController {
         print("will display json")
         let start = Date().timeIntervalSince1970
         
-        previewView.preview(json, initialState: .folded) {
+        previewView.preview(json, initialState: .folded) { _ in
             let end = Date().timeIntervalSince1970
             print("did display json at: \(end - start)")
         }